# jobnow-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "JobNow API",
        "description": "API documentation for JobNow - Moroccan Job Portal Platform. This API provides endpoints for job seekers and companies to manage job offers, applications, and user profiles.",
        "contact": {
            "name": "JobNow Support",
            "email": "support@jobnow.ma"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.jobnow.ma",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/candidat/saved-jobs": {
            "get": {
                "tags": [
                    "Saved Jobs"
                ],
                "summary": "Get all saved jobs for authenticated candidat",
                "description": "Returns list of jobs saved by the candidat",
                "operationId": "d51bef1292e55d59c2bd3d7b02368f16",
                "responses": {
                    "200": {
                        "description": "Saved jobs retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/candidat/saved-jobs/{offreId}": {
            "post": {
                "tags": [
                    "Saved Jobs"
                ],
                "summary": "Save a job offer",
                "description": "Add a job offer to candidat's saved jobs",
                "operationId": "657d0e6bc59ece715e69fff76f07501e",
                "parameters": [
                    {
                        "name": "offreId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Job saved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Offre sauvegardée avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Job offer not found"
                    },
                    "409": {
                        "description": "Job already saved"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Saved Jobs"
                ],
                "summary": "Remove a saved job",
                "description": "Remove a job offer from candidat's saved jobs",
                "operationId": "43b07698d531c34bf65bab08cb540518",
                "parameters": [
                    {
                        "name": "offreId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job removed from saved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Offre retirée des favoris."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Saved job not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/candidat/saved-jobs/check/{offreId}": {
            "get": {
                "tags": [
                    "Saved Jobs"
                ],
                "summary": "Check if job is saved",
                "description": "Check if a specific job is saved by the candidat",
                "operationId": "58586d48f839a22246b9de0c1f0a4803",
                "parameters": [
                    {
                        "name": "offreId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Check result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "is_saved": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}